requirements.txt
colorama
pyyaml
requests
tabulate
tqdm
